infinite loopers cancer detection/requirements.txt
Flask==3.0.0
google-generativeai==0.4.0
Pillow==10.2.0
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0
